move prompt to io/
